extract Dropdown(Theme), to single component
